Rename, use translation file for text, simplified
Ticket: #35
