+++ src/app/Main/Body/QueryPage/ExampleQueriesDialog/ExampleQueries.constants.ts
@@ -6,8 +6,10 @@ const defaultQuery2 = "?microorganism=Salmonella+spp.&samplingYear=2015";
 const defaultQuery3 =
     "?microorganism=CARBA-E.+coli&row=microorganism&column=samplingYear";
 
-export const defaultQueriesLists = {
+export type QueryCategories = "animal" | "food" | "animalFeed";
+
+export const exampleQueriesLists: Record<QueryCategories, string[]> = {
     animal: [defaultQuery1, defaultQuery2, defaultQuery3],
     food: [defaultQuery1, defaultQuery2, defaultQuery3],
-    animalFood: [defaultQuery1, defaultQuery2, defaultQuery3],
+    animalFeed: [defaultQuery1, defaultQuery2, defaultQuery3],
 };
