+++ src/locales/en/ExampleQueries.json
@@ -1,0 +1,59 @@
+{
+  "Content": {
+    "Title": "Example queries",
+    "Text": "Zeitverlauf von Antibiotikaresistenz von Isolaten von Salmonella spp. aus Legehennen, die im Rahmen der Bekämpfungsprogrammen beim Geflügel gewonnen wurden"
+  },
+  "Subtitles": {
+    "animal": "Animal",
+    "food": "Food",
+    "animalFeed": "Animal feed"
+  },
+  "Queries": {
+    "animal": {
+      "Query0": {
+        "Title": "Query0",
+        "Text": "Describe Query0"
+      },
+      "Query1": {
+        "Title": "Query1",
+        "Text": "Describe Query1"
+      },
+      "Query2": {
+        "Title": "Query2",
+        "Text": "Describe Query2"
+      }
+    },
+    "food": {
+      "Query0": {
+        "Title": "Query0",
+        "Text": "Describe Query0"
+      },
+      "Query1": {
+        "Title": "Query1",
+        "Text": "Describe Query1"
+      },
+      "Query2": {
+        "Title": "Query2",
+        "Text": "Describe Query2"
+      }
+    },
+    "animalFeed": {
+      "Query0": {
+        "Title": "Query0",
+        "Text": "Describe Query0"
+      },
+      "Query1": {
+        "Title": "Query1",
+        "Text": "Describe Query1"
+      },
+      "Query2": {
+        "Title": "Query2",
+        "Text": "Describe Query2"
+      }
+    }
+  },
+  "Button": {
+    "Submit": "Auswählen",
+    "Cancel": "Abbrechen"
+  }
+}
